fix(security): restringe Régua de Emoções a usuários autorizados (servidor)
Módulo não liberado: autorização passa a ser server-authoritative via
can_view_emotion_ruler (ou sudo), nunca TOTEM/desativado.

- access-control: novo helper canViewEmotionRuler (reutilizável)
- emotion-ruler router: gate em getActive, createResponse, registerAccess e
  registerDismissal (createResponse antes não tinha NENHUMA checagem)
- /forms/emotion-ruler: page vira server component com guard + redirect; UI
  extraída para EmotionRulerResponseForm (client)
- sidebar/quick-access (routes.ts): item gated por permissão em vez de "novidades"
- Bump de versão para 1.36.2 (MMP - patch)
