+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "elo",
-  "version": "1.36.1",
+  "version": "1.36.2",
   "private": true,
   "type": "module",
   "scripts": {
